Remove windows include from threading header
